buildable package for bochord
