+++ bochord.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+# Run bochord in development
+poetry run ./bochord.py "$@"
